% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisec.R
\name{bisec}
\alias{bisec}
\title{Bisection algorithm}
\usage{
bisec(delta,dea1,dea2,upbd,lowbd)
}
\arguments{
\item{delta}{}

\item{dea1}{}

\item{dea2}{}

\item{upbd}{}

\item{lowbd}{}
}
\value{
a2
}
\description{
Running bisection algorithm to search for a2, the minimizer of (log((a2)^dea1*(1-a2)^dea2-delta))^2
Upbd and lowbd are the upper and lower bounds.
}

