% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{df}
\alias{df}
\title{JAMA Breast cancer}
\usage{
data(df)
}
\description{
A dataset containing predictions for chemo-censitivity and pathological response 
from Hatzis (2011)
}
\details{
\itemize{
  \item validate: Validation status
  \item drfs: Censoring status; 0=censored, 1=event.
  \item drfs.time: Time to event or censoring
  \item er.status: ER status, P=positive, N=negative
  \item chemo.pred: Prediction for chemo sensitivity from the ACES predictor, sensitive or insensitive
  \item pre.N: Prediction of nodal status
  \item pCR.RD: pathological complete response (pCR) or residual disease (RD) 
  \item pre.grade: prediction of tumor grade
  \item pre.T: T stage prediction
  \item dlda30: DLDA30 prediction for the pathological response.
  
}
}
\references{
[1] Hatzis, C., Pusztai, L., Valero, V., Booser, D. J., Esserman, L., Lluch, A., et al. (2011). A genomic predictor of response and survival following taxane-anthracycline chemotherapy for invasive breast cancer. The Journal of the American Medical Association 305, 1873–1881.
}
\keyword{datasets}
