% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km}
\alias{km}
\title{Plot a Kaplan Meier curve}
\usage{
km(time, censor, plotcens)
}
\arguments{
\item{time}{time of observed event}

\item{censor}{a vector indicating censored or not at the given times, 0 = censored; 1 = uncensored}

\item{plotcens}{0: add censored data to the output curve

                1: don't add censored data to the output curve}
}
\value{
A Kaplan Meier curve
}
\description{
The function...
}

