% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_log2.R
\name{km_log2}
\alias{km_log2}
\title{Plot a red Kaplan Meier curve in log scale}
\usage{
km_log2(time, censor, plotcens)
}
\arguments{
\item{time}{time of observed event}

\item{censor}{a vector indicating censored or not at the given times, 0 = censored; 1 = uncensored}

\item{plotcens}{0: add censored data to the output curve

                1: don't add censored data to the output curve}
}
\value{
A red Kaplan Meier curve in log scale
}
\description{
The function...
}

