% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_red.R
\name{km_red}
\alias{km_red}
\title{Plot a Kaplan Meier curve in red}
\usage{
km_red(time, censor, plotcens)
}
\arguments{
\item{time}{time of observed event}

\item{censor}{a vector indicating censored or not at the given times, 0 = censored; 1 = uncensored}

\item{plotcens}{0: add censored data to the output curve

                1: don't add censored data to the output curve}
}
\value{
A red Kaplan Meier curve
}
\description{
The function plots a Kaplan Meier curve in red
}
\examples{
t1 <- c(2,3,4,5.5,7,10,12,15)
c1 <- c(0,0,1,0,0,1,0,0)
km_red(t1,c1,0)
}
