% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_redsolid.R
\name{km_redsolid}
\alias{km_redsolid}
\title{Plot a Kaplan Meier curve in red solid line}
\usage{
km_redsolid(time, censor, plotcens)
}
\arguments{
\item{time}{time of observed event}

\item{censor}{a vector indicating censored or not at the given times, 0 = censored; 1 = uncensored}

\item{plotcens}{0: add censored data to the output curve

                1: don't add censored data to the output curve}
}
\value{
A red solid Kaplan Meier curve
}
\description{
The function...
}
\examples{

}

