% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot.R
\name{kmplot}
\alias{kmplot}
\title{Obtain values for Kaplan-Meier plotting}
\usage{
kmplot(x, alpha)
}
\arguments{
\item{x}{Nx2 data matrix,first columen represents survival time of the i-th subject, second column represents censored flag (0 if not censored, 1 if censored)}

\item{alpha}{significance level}
}
\value{
Values used for Kaplan-Meier plotting
}
\description{
Obtain values for Kaplan-Meier plotting
}

