% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrank_Yates228.R
\name{logrank_Yates228}
\alias{logrank_Yates228}
\title{Obtain the logrank test}
\usage{
logrank_Yates228(x1, x2, Alpha)
}
\arguments{
\item{x1}{Nx2 data matrix,first columen represents survival time of the i-th subject, second column represents censored flag (0 if not censored, 1 if censored)}

\item{x2}{Nx2 data matrix,first columen represents survival time of the i-th subject, second column represents censored flag (0 if not censored, 1 if censored)}

\item{Alpha}{Significance level(optional, default is 0.05)}
}
\value{
Log-rank statistics
}
\description{
Comparing survival curves of two groups using the log rank test
Comparison of two survival curves can be done using a statistical
hypothesis test called the log rank test. It is used to test the null
hypothesis that there is no difference between the population survival
curves (i.e. the probability of an event occurring at any time point is
the same for each population).
}
\examples{
logrank_Yates228(x1,x2,0.05)
logrank_Yates228(x1,x2)
}

