% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{loopcut_onestep_data}
\alias{loopcut_onestep_data}
\title{Example data for loopcut_onestep}
\usage{
data(loopcut_onestep_data)
}
\description{
Example data for loopcut_onestep
}
\keyword{datasets}
