% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{loopcuts_t_c}
\alias{loopcuts_t_c}
\title{Example data for loopcut_times_censoring}
\usage{
data(loopcuts_t_c)
}
\description{
Example data for loopcut_times_censoring
}
\keyword{datasets}
