% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{loopcuts_umbrella_cuttimes_mono}
\alias{loopcuts_umbrella_cuttimes_mono}
\title{Example data for loopcut_umbrella}
\usage{
data(loopcuts_umbrella_cuttimes_mono)
}
\description{
Example data for loopcut_umbrella
}
\keyword{datasets}
