% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pava_dfr.R
\name{pava_dfr}
\alias{pava_dfr}
\title{PAVA order restriction under decreasing failure rate (DFR)}
\usage{
pava_dfr(time_die,ttot,deaths)
}
\arguments{
\item{time_die}{event times}

\item{ttot}{the total time on test (ttot) corresponding to the event times}

\item{deaths}{the number of deaths at each event time}
}
\value{
time2: the event times after PAVA
ttot2: the corresponding ttot
deaths2 the corresponding number of deaths
}
\description{
This function imposes the PAVA DFR order restriction by eliminating change-points violating the restriction
}
\examples{
data(pava_dfrd)
t_d = pava_dfrd[,1]
t = pava_dfrd[,2]
d = pava_dfrd[,3]
pava_dfr(t_d, t, d)
}
