% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexeest_p.R
\name{pexeest_p}
\alias{pexeest_p}
\title{pexeest_p}
\usage{
pexeest_p(times, cens, tchange, tx)
}
\arguments{
\item{times}{}

\item{cens}{}

\item{tchange}{}

\item{tx}{}
}
\value{
pchange
}
\description{
a function produces the the quantile estimate at tx, 
when a piecewise exponential distribution is fitted to 
(times,cens) cens = 0 for censored, cens = 1 for uncensored.
}

