% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{pexeest_times_censoring}
\alias{pexeest_times_censoring}
\title{Example data for pexeest_times_censoring}
\usage{
data(pexeest_times_censoring)
}
\description{
Example data for pexeest_times_censoring
}
\keyword{datasets}
