% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{simple}
\alias{simple}
\title{None Small Cell Lung cancer data}
\usage{
data(simple)
}
\description{
A dataset non-small-cell lung cancer trial data from Simon et al. (2011)
}
\details{
\itemize{
  \item first column - censor : censoring status; 0=censored, 1=event.
  \item second column - times : time to event
}
}
\references{
[1] Simon GR, Extermann M, Chiappori A, Williams C, Begum M, Haura RKE, Ismail-Khan R, Schell M, Antonia SJ, Bepler G. Phase 2 trial of docetaxel and gefitinib in the first-line treatment of patients with advanced stage non-small cell lung cancer (NSCLC) who are 70 years of age or older. Cancer 2008; 112:2021–2029.
}
\keyword{datasets}
