% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totaltest.R
\name{totaltest}
\alias{totaltest}
\title{Computes total time on test}
\usage{
totaltest(time,censor)
}
\arguments{
\item{time}{}

\item{censor}{}
}
\value{
time_die
}
\description{
Computes total time on test
}

