% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{data2}
\alias{data2}
\title{RPEXE_fitting}
\usage{
data(data2)
}
\description{
A dataset containing ....
}
\details{
\itemize{
  \item first column. times
  \item second column. censor
  \item group
}
}
\keyword{datasets}
