% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{df}
\alias{df}
\title{JAMA Breast cancer}
\usage{
data(df)
}
\description{
A dataset containing ....
}
\details{
\itemize{
  \item validate: ... 
  \item drfs: .... 
  \item drfs.time: ...
  \item group
}
}
\keyword{datasets}
