% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamllik.R
\name{gamllik}
\alias{gamllik}
\title{Computing the log likelihood from the gamma distribution}
\usage{
gamllik(structtime,structttot,structdeaths,time_die,ttot,deaths)
}
\arguments{
\item{structtime}{times under restriction}

\item{structttot}{ttots between each time point and the previous
time point (or 0) under restriction}

\item{structdeaths}{number of deaths corresponding to struct.ttot}

\item{time_die}{all possible times to make the cuts}

\item{ttot}{ttots corresponding to time_die}

\item{deaths}{dealthes corresponding to ttot}
}
\value{
log of the likelihood
}
\description{
A function computing the log likelihood from the gamma distribution under
      an order restriction reduction
}
\examples{

}

