% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_blacksolid.R
\name{km_blacksolid}
\alias{km_blacksolid}
\title{Kaplan-Meier curve}
\usage{
km_blacksolid(time, censor, plotcens)
}
\arguments{
\item{time}{a vector of event or censoring time}

\item{censor}{a vector indicating censoring: 0 = censored; 1 = uncensored}

\item{plotcens}{0: don't add censored data symbol to the output curve
1: add censored data symbol to the output curve}
}
\value{
Kaplan-Meier curve only
}
\description{
This function plots the Kaplan-Meier curve without returning outputs
}
\examples{
t1 <- c(2,3,4,5.5,7,10,12,15)
c1 <- c(0,0,1,0,0,1,0,0)
km_blacksolid(t1,c1,0)

}
