% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_combine.R
\name{km_combine}
\alias{km_combine}
\title{Comparing two Kaplan Meier curves in one plot}
\usage{
km_combine(x1, x2)
}
\arguments{
\item{x1}{}

\item{x2}{}
}
\value{
A combined Kaplan Meier curve
}
\description{
The function...
}

