% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pava_dfr.R
\name{pava_dfr}
\alias{pava_dfr}
\title{Pava_dfr}
\usage{
pava_dfr(time_die,ttot,deaths)
}
\arguments{
\item{time_die}{}

\item{ttot}{}

\item{deaths}{}
}
\value{
time2: the merged time_die
ttot2: the merged ttot
deaths2 the merged deaths
}
\description{
Merge certain entries to make the sequence of ttot to be non decreasing
}

