% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pava_ifr.R
\name{pava_ifr}
\alias{pava_ifr}
\title{Pava_ifr}
\usage{
pava_ifr(time_die,ttot,deaths)
}
\arguments{
\item{time_die}{}

\item{ttot}{}

\item{deaths}{}
}
\value{
time2: the merged time_die
ttot2: the merged ttot
deaths2 the merged deaths
}
\description{
Merge certain entries to make the sequence of ttot to be non increasing
}

