% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{package}
\name{RPEXE.RPEXT}
\alias{RPEXE.RPEXT}
\alias{RPEXE.RPEXT-package}
\title{RPEXE.RPEXT}
\description{
The package allows you to .....
}
\details{
This reduced piecewise exponential survival software implements the likelihood 
ratio test procedure in Han, Schell, and Kim (2009). Inputs to the program can be either 
times when events/censoring occur or the vectors of total time on test and the number of 
events. Outputs of the programs are times of events and the corresponding p- values. The 
order of times and p-values is determined by a backward elimination procedure. Details 
about the model and implementation are given in Han, Schell, and Kim (2009). This program
can run in R version 12 and above.
}
\references{
Han, G., Schell, M. J., and Kim, J. (2009) Improved survival modeling using a piece- wise exponential approach
}

