% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXEv1_2.R
\name{RPEXEv1_2}
\alias{RPEXEv1_2}
\title{RPEXE}
\usage{
RPEXEv1_2(eventtime,censor,cuttime, trend,criticalps)
}
\arguments{
\item{eventtime}{A sequence of times where the events occur}

\item{censor}{A sequence of dichotomous values indicating censored or not (0=censored and 1=not censored)}

\item{cuttime}{A vector of unique, sorted, possible times to make the cuts. Default is sorted (from small to
large) event times
Default = 'EventTime'}

\item{trend}{An input having indicating the monotonicity assumption
-- 0: no monotonic assumption

-- 1: failure rate is decreasing over time

-- 2: failure rate is increasing over time

-- 3: monotonic failure rate

-- 4: failure rate is increasing and then decreasing

-- 5: failure rate is decreasing and then increasing

-- 6: failure rate is increasing and then decreasing with
      the peak removed first
      
-- 7: failure rate is decreasing and then increasing with
      the peak removed first
      
Default == 0}

\item{criticalps}{The critical (naive) p-value cutoff where all p-values 
in the backward elimination that are lower than this 
will be regarded as being significant. The prediction of 
the survival probability will be made on 100 equally
spaced time points within the range of the event times 
based on the piecewise exponential estimate determined by 
all the changepoints. 
Default == -1 (equivalent to NA).}
}
\value{
times: times to make the cuts
pvalues: pvalues correspond to the times
times_c: critical times to make the cuts
pvalues_c: critical p-values that are smaller than the 
trend: trend information
struct:  structure information for multiple order restrictions
changet:  change point in time for umbrella alternatives.
}
\description{
RPEXE
}

