% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPEXE.RPEXT.R
\docType{data}
\name{simple}
\alias{simple}
\title{None Small Cell Lung cancer data}
\usage{
data(simple)
}
\description{
A dataset containing ....
}
\details{
\itemize{
  \item validate: ... 
  \item drfs: .... 
  \item drfs.time: ...
  \item group
}
}
\keyword{datasets}
