% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totaltest.R
\name{totaltest}
\alias{totaltest}
\title{total time on test}
\usage{
totaltest(time,censor)
}
\arguments{
\item{time}{event/censoring times}

\item{censor}{censoring status}
}
\value{
time_die time points where events occur (in ascending order)
ttot total time on test corresponding to each time point in "time_die"
deaths number of death corresponding to each time point in "time_die"
}
\description{
Function 'totaltest' computes total-time-on-test.
}
\examples{
t1 <- c(2,3,4,5.5,7,10,12,15)
c1 <- c(0,0,1,0,0,1,0,0)
totaltest(t1,c1)
}
