% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbrella.R
\name{umbrella}
\alias{umbrella}
\title{umbrella}
\usage{
umbrella(time_die,ttot,deaths,indi)
}
\arguments{
\item{time_die}{a sequence of times where deaths happened.}

\item{ttot}{the total time on test between each time point
and the previous time point (or 0).}

\item{deaths}{the number of deaths at each time point.}

\item{indi}{an indicator

               indi = 0: monotonic failure rate (either decrease or increase)
               
               indi = 1: denoting the failure rate increase then decrease
               
               indi = 2: denoting the failure rate decrease then
                          increase}
}
\description{
Using the umbrella alternative to merge certain entries to make
   the sequence of ttot/deaths to increase then decrease
}

